requirements.txt
flask
gunicorn

line-bot-sdk

requests
BeautifulSoup4
lxml